+++ common/src/main/kotlin/com/cobblemon/mod/common/api/events/CobblemonEvents.kt
@@ -38,6 +38,7 @@ import com.cobblemon.mod.common.api.events.pokemon.interaction.HeldItemUpdatedEv
 import com.cobblemon.mod.common.api.events.pokemon.interaction.PokemonInteractionGUICreationEvent
 import com.cobblemon.mod.common.api.events.starter.StarterChosenEvent
 import com.cobblemon.mod.common.api.events.storage.ReleasePokemonEvent
+import com.cobblemon.mod.common.api.events.storage.StorePokemonEvent
 import com.cobblemon.mod.common.api.events.world.BigRootPropagatedEvent
 import com.cobblemon.mod.common.api.reactive.CancelableObservable
 import com.cobblemon.mod.common.api.reactive.EventObservable
@@ -174,4 +175,9 @@ object CobblemonEvents {
     val POKEMON_GAINED = EventObservable<PokemonGainedEvent>()
     @JvmField
     val POKEMON_SEEN = EventObservable<PokemonSeenEvent>()
+
+    @JvmField
+    val COLLECT_EGG = EventObservable<CollectEggEvent>()
+    @JvmField
+    val HATCH_EGG = EventObservable<HatchEggEvent>()
 }